requirements.txt
Flask == 2.0.0
gunicorn==20.0.4
requests
pyyaml